Moved setup of playwright to seperate file
Seperated the setup of the WebApplicationFactory to a seperate fil to keep the test files clean.
